# Add hash-join execution for left-outer-joins