--- a/server/src/main/java/io/crate/breaker/TypedRowAccounting.java
+++ b/server/src/main/java/io/crate/breaker/TypedRowAccounting.java
@@ -26,7 +26,7 @@ import java.util.List;
 import io.crate.data.Row;
 import io.crate.data.breaker.RamAccounting;
 import io.crate.data.breaker.RowAccounting;
-import io.crate.execution.engine.join.HashInnerJoinBatchIterator;
+import io.crate.execution.engine.join.HashJoinBatchIterator;
 import io.crate.types.DataType;
 
 /**
@@ -52,7 +52,7 @@ public class TypedRowAccounting implements RowAccounting<Row> {
      *
      * @param columnTypes     The column types to use for size estimation
      * @param ramAccounting   {@link RamAccounting} implementing the CircuitBreaker logic
-     * @param extraSizePerRow Extra size that need to be calculated per row. E.g. {@link HashInnerJoinBatchIterator}
+     * @param extraSizePerRow Extra size that need to be calculated per row. E.g. {@link HashJoinBatchIterator}
      *                        might instantiate an ArrayList per row used for the internal hash->row buffer
      */
     public TypedRowAccounting(List<? extends DataType<?>> columnTypes,