… docs/appendices/release-notes/5.10.0.rst
… docs/general/dql/joins.rst
… docs/sql/statements/explain.rst
… server/src/test/java/io/crate/execution/engine/join/HashJoinBatchIteratorBehaviouralTest.java
… server/src/test/java/io/crate/execution/engine/join/HashJoinBatchIteratorMemoryTest.java
… server/src/test/java/io/crate/execution/engine/join/LeftOuterHashJoinBatchIteratorTest.java
… server/src/test/java/io/crate/integrationtests/JoinIntegrationTest.java
… server/src/test/java/io/crate/integrationtests/LookupJoinIntegrationTest.java
… server/src/test/java/io/crate/integrationtests/ShowIntegrationTest.java
… server/src/test/java/io/crate/planner/ExplainPlannerTest.java
… server/src/test/java/io/crate/planner/SelectPlannerTest.java
… server/src/test/java/io/crate/planner/ViewPlannerTest.java
… server/src/test/java/io/crate/planner/node/dql/JoinPhaseTest.java
… server/src/test/java/io/crate/planner/operators/EquiJoinDetectorTest.java
… server/src/test/java/io/crate/planner/operators/JoinTest.java
… server/src/test/java/io/crate/planner/operators/LogicalPlannerTest.java
… server/src/test/java/io/crate/planner/operators/OuterJoinRewriteTest.java
… server/src/test/java/io/crate/planner/operators/PushDownTest.java
… server/src/test/java/io/crate/planner/operators/RelationNamesInLogicalPlanTest.java
… server/src/test/java/io/crate/planner/operators/SelectDistinctLogicalPlannerTest.java
… server/src/test/java/io/crate/planner/optimizer/costs/PlanStatsTest.java