## Summary of the changes / Why this improves CrateDB

This implements hash-join execution for left-outer-joins according to https://github.com/crate/crate/issues/15147 by the following steps:

- Extend `HashInnerJoinBatchIterator` to `HashJoinBatchIterator` with an additional option to emit non-matching join pairs with null values to support left-outer-join behaviour 
- Extend the planning pipeline to support `Left` Join-Type for Hash-Joins


## Performance Improvements:
This is a comparison between the hash-join and nested-loop join execution on a table having 50 % matches and 50 % non-matches with different sizes running each time 100 iterations:
```
select count(*) from t1 left join t2 on t1.id = t2.id
```


| Table sizes | Nestedloop (ms mean) | Hash-Join (ms mean)| 
| ------------- | ------------- | ------------- | 
| 10 / 10  | 2.045 ± 1.229 |  0.555 ± 0.110  | 
| 100 / 100  | 4.455 ± 0.271  |  0.438 ± 0.028 | 
| 1000 / 1000  | 169.494 ± 0.286  | 1.058 ± 0.018  | 
| 10.000 / 10.000  | 21919.926 ± 653.076  | 7.955 ± 0.050  | 
| 100.000 / 100.000  | n/a  | 178.404 ± 4.188  | 

Fixes https://github.com/crate/crate/issues/15147