@mkleen 2024-10-24
There seems to be no overhead on the inner-hash-join according to the jmh bench:


this branch:

```
Result "io.crate.data.join.RowsBatchIteratorBenchmark.measureConsumeHashInnerJoin":
  2.646 ±(99.9%) 0.011 ms/op [Average]
  (min, avg, max) = (2.628, 2.646, 2.683), stdev = 0.014
  CI (99.9%): [2.636, 2.657] (assumes normal distribution)

Result "io.crate.data.join.RowsBatchIteratorBenchmark.measureConsumeHashInnerJoinWithHashCollisions":
  0.323 ±(99.9%) 0.003 ms/op [Average]
  (min, avg, max) = (0.318, 0.323, 0.331), stdev = 0.004
  CI (99.9%): [0.320, 0.326] (assumes normal distribution)
  
Benchmark                                                                 Mode  Cnt  Score   Error  Units
RowsBatchIteratorBenchmark.measureConsumeHashInnerJoin                    avgt   25  2.646 ± 0.011  ms/op
RowsBatchIteratorBenchmark.measureConsumeHashInnerJoinWithHashCollisions  avgt   25  0.323 ± 0.003  ms/op
```

master:

```
Result "io.crate.data.join.RowsBatchIteratorBenchmark.measureConsumeHashInnerJoin":
  2.679 ±(99.9%) 0.019 ms/op [Average]
  (min, avg, max) = (2.645, 2.679, 2.716), stdev = 0.025
  CI (99.9%): [2.660, 2.698] (assumes normal distribution)

Result "io.crate.data.join.RowsBatchIteratorBenchmark.measureConsumeHashInnerJoinWithHashCollisions":
  0.302 ±(99.9%) 0.002 ms/op [Average]
  (min, avg, max) = (0.297, 0.302, 0.311), stdev = 0.003
  CI (99.9%): [0.299, 0.304] (assumes normal distribution)
  
Benchmark                                                                 Mode  Cnt  Score   Error  Units
RowsBatchIteratorBenchmark.measureConsumeHashInnerJoin                    avgt   25  2.679 ± 0.019  ms/op
RowsBatchIteratorBenchmark.measureConsumeHashInnerJoinWithHashCollisions  avgt   25  0.302 ± 0.002  ms/op
```