@matriv 2024-10-24
Please remove file `.pom.xml.swp`.