@matriv 2024-10-24
> There seems to be no overhead on the inner-hash-join according to the jmh bench:
great, thank you!

> I'd say to add a JMH benchmark as well.
Maybe you can do that already, but can also be done as a followup, as you like

> I think it'll make sense to run end-to-end benchmarks for different sizes of tables to also see if for small sizes, NL could be faster than HashJoin.
Do you have any metrics for this? I think it would be nice to have an indication before merging the PR

Otherwise it look great, nice stuff!
Please give me just a bit more time, to take a closer look in the iterator (debug to understand it thoroughly), and I think we're good to go!