@mkleen 2024-10-25
> Do you have any metrics for this? I think it would be nice to have an indication before merging the PR

It's in the description of this pr.